[kpm][nl] Work on AApiAiParseUserTextAgent
